book-library-mobile/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
beautifulsoup4==4.12.3
